chore: renamed node2 to node
